requirements.txt
pyttsx3
SpeechRecognition
pyaudio
google-generativeai
python-dotenv